Moved authenticate() to LoginModel. Gets called in the controller.
